+++ examples/map_events.rs
@@ -19,7 +19,7 @@ fn main() {
         .add_observer(map_created)
         .add_observer(layer_created)
         .add_observer(object_created)
-        .add_observer(special_tile_created)
+        .add_observer(tile_created)
         // Add our systems and run the app!
         .add_systems(Startup, startup)
         .run();
@@ -36,12 +36,12 @@ fn map_created(
     map_asset: Res<Assets<TiledMap>>,
 ) {
     // We can either access the map components
-    if let Ok(name) = q_map.get(trigger.event().map) {
+    if let Ok(name) = q_map.get(trigger.event().entity) {
         info!("Received TiledMapCreated event for map '{}'", name);
     }
 
-    // Or directly the underneath Tiled Map structure
-    let map = trigger.event().map(&map_asset);
+    // Or directly the underneath tiled Map data
+    let map = trigger.event().get_map(&map_asset);
     info!("Loaded map: {:?}", map);
 }
 
@@ -51,13 +51,13 @@ fn layer_created(
     map_asset: Res<Assets<TiledMap>>,
 ) {
     // We can either access the layer components
-    if let Ok(name) = q_layer.get(trigger.event().layer) {
+    if let Ok(name) = q_layer.get(trigger.event().entity) {
         info!("Received TiledLayerCreated event for layer '{}'", name);
     }
 
     // Or directly the underneath Map or Layer structures
-    let _map = trigger.event().map(&map_asset);
-    let layer = trigger.event().layer(&map_asset);
+    let _map = trigger.event().map.get_map(&map_asset);
+    let layer = trigger.event().get_layer(&map_asset);
     info!("Loaded layer: {:?}", layer);
 }
 
@@ -67,30 +67,35 @@ fn object_created(
     map_asset: Res<Assets<TiledMap>>,
 ) {
     // We can either access the object components
-    if let Ok(name) = q_object.get(trigger.event().object) {
+    if let Ok(name) = q_object.get(trigger.event().entity) {
         info!("Received TiledObjectCreated event for object '{}'", name);
     }
 
     // Or directly the underneath Map, Layer or Object structures
-    let _map = trigger.event().map(&map_asset);
-    let _layer = trigger.event().layer(&map_asset);
-    let object = trigger.event().object(&map_asset);
+    let _map = trigger.event().layer.map.get_map(&map_asset);
+    let _layer = trigger.event().layer.get_layer(&map_asset);
+    let object = trigger.event().get_object(&map_asset);
     info!("Loaded object: {:?}", object);
 }
 
-fn special_tile_created(
-    trigger: Trigger<TiledSpecialTileCreated>,
+// Note that only tiles which have custom properties will trigger this event,
+// even without the crate `user_properties` feature.
+// The Debug implementation for tiled::LayerTile does not display the actual
+// content of the properties field but it's there, go check the tileset if you
+// don't believe me :)
+fn tile_created(
+    trigger: Trigger<TiledTileCreated>,
     q_tile: Query<&Name, With<TiledMapTile>>,
     map_asset: Res<Assets<TiledMap>>,
 ) {
     // We can either access the tile components
-    if let Ok(name) = q_tile.get(trigger.event().tile) {
-        info!("Received TiledSpecialTileCreated event for tile '{}'", name);
+    if let Ok(name) = q_tile.get(trigger.event().entity) {
+        info!("Received TiledTileCreated event for tile '{}'", name);
     }
 
     // Or directly the underneath Map and Layer structures
-    let _map = trigger.event().map(&map_asset);
-    let _layer = trigger.event().layer(&map_asset);
-    let tile = trigger.event().tile(&map_asset);
+    let _map = trigger.event().layer.map.get_map(&map_asset);
+    let _layer = trigger.event().layer.get_layer(&map_asset);
+    let tile = trigger.event().get_tile(&map_asset);
     info!("Loaded tile: {:?}", tile);
 }